+++ CumulusMX/WeatherStation.cs
@@ -413,6 +413,7 @@ namespace CumulusMX
 			versionCheckTime = new DateTime(1, 1, 1, Program.RandGenerator.Next(0, 23), Program.RandGenerator.Next(0, 59), 0, DateTimeKind.Local);
 
 			SensorReception = [];
+			SensorRssi = [];
 
 			DayFileQuery = new QueryDayFile(RecentDataDb);
 		}
@@ -8429,6 +8430,7 @@ namespace CumulusMX
 		private bool dayfileReloading;
 
 		public static Dictionary<string, byte> SensorReception { get; set; }
+		public static Dictionary<string, int> SensorRssi { get; set; }
 
 		public void WriteYesterdayFile(DateTime logdate)
 		{